added text compression functions
