--- test.yml
@@ -1,3 +1,0 @@
-a: 3
-b: 2
-c: 4
