+++ weather.py
@@ -165,249 +165,10 @@ def check_type(orig_path):
     quit()
 
 #####################################################################
-# B u i l d   R G B   s c a l e
-#####################################################################
-def scale_build_int(raw):
-    # use a slice of RGB values to return an array of BGR integers
-    scale_out = []
-    s = 0
-    r_last = np.astype(np.array([0,0,0]),np.uint8)
-    for r in raw:
-        # first, filter out black and white
-        if not all(r > 250) and not all(r < 5):
-            # next, check if the RGB value matches the last RGB
-            if sum(np.abs(np.subtract(r_last,r))) >= 5:
-                for BGR in r:
-                    s = s * 256 + int(BGR)
-                scale_out.append(s)
-                s = 0
-                r_last = r.copy()
-    return scale_out
-def scale_build_RGB(scale_from_config):
-    # Fix: get rid of this once a yml file is used for configs
-    # use an array of BGR intgers to return the BGR values as an array
-    scale = []
-    for sc in scale_from_config:
-        scale_element = [0,0,0]
-        scale_element[2] = sc % 256
-        sc = int((sc - scale_element[0]) / 256)
-        scale_element[1] = sc % 256
-        sc = int((sc - scale_element[1]) / 256)
-        scale_element[0] = sc
-        scale.append(scale_element)
-    return np.astype(np.array(scale),np.uint8)
-
-# Fix: shift all these functions to their own file plot.py
-#####################################################################
-# C r e a t e   a   p l o t   o f   s c a l a r   v a l u e s 
-#####################################################################
-def get_tblr(image):
-    def get_tblr_bounds(mask_bounds):
-        start, end = len(mask_bounds[:,0]), 0
-        found = False
-        temp = 0
-        for igtb in range(len(mask_bounds[:,0])):
-            if sum(mask_bounds[igtb,:]) > 0 and not found:
-                found = True
-                temp = igtb
-            if sum(mask_bounds[igtb,:]) == 0 and found:
-                found = False
-                if igtb - temp > end - start:
-                    end = igtb - 1
-                    start = temp
-        if found:
-            if igtb - temp > end - start:
-                end = igtb - 1
-                start = temp
-        return [start, end]
-    #create a mask of non-black/non-white pixels
-    mask = np.zeros_like(image[:,:,0]).astype(bool)
-    for igt in range(3):
-        mask = mask + np.multiply(image[:,:,igt] > 15, image[:,:,igt] < 240)
-    t, b, l, r = 0, len(mask[:,0]), 0, len(mask[0,:])
-    for jgt in range (3):
-        [t,b] = get_tblr_bounds(mask[t:b,l:r])
-        [l,r] = get_tblr_bounds(mask[t:b,l:r].T)
-    return [t,b,l,r]
-def plot_scale(image):
-    arr = image.copy()
-    # create a blank canvas and import RGB values for the scale
-    plot_out = np.zeros_like(image[:,:,0]).astype(int)
-    # build a scalar plot for each level of the scale
-    for i in range(len(scale)):
-        # build a bool array based on what is still blank in the output
-        plot_bool = plot_out == 0
-        # remove all pixels which do not match the given BGR value
-        for j in range(3): 
-            plot_bool = np.multiply(plot_bool, abs(arr[:,:,j]-scale[i,j]) < 5)
-        plot_out = plot_out + (plot_bool.astype(int) * (i + 1))
-    return plot_out
-def plot_smooth(arr,repeat):
-    # function to smooth out the colors and crop out the borders
-    arr_out = np.copy(arr)
-    # use queen smoothing to get rid of non-colored lines
-    for r in range(repeat):
-        arr_tmp = np.pad(
-            arr_out[1:arr_out.shape[0]-1,1:arr_out.shape[1]-1],
-            (1,),
-            'constant',constant_values=0
-        )
-        arr_add = np.zeros_like(arr_out).astype(np.float64)
-        arr_cnt = np.zeros_like(arr_out).astype(np.float64)
-
-        for (i, j) in [(1,0),(1,1),(-1,0),(-1,0),(-1,1),(-1,1),(1,0)]:
-            arr_tmp = np.roll(arr_tmp, i, axis = j)
-            arr_add = arr_add + arr_tmp
-            arr_cnt = (arr_cnt 
-                       + np.array(arr_tmp != 0).astype(type(arr_cnt[0,0])))
-        arr_add = np.divide(
-            arr_add, arr_cnt, 
-            out = np.zeros_like(arr_add), 
-            where=arr_cnt!=0
-        )
-        if test_mode: 
-            cv.imwrite('./debug/test_add_' 
-                       + str(r) 
-                       + '.jpg', 
-                       10 * arr_add)
-        if test_mode: 
-            cv.imwrite('./debug/test_cnt_' 
-                       + str(r) 
-                       + '.jpg', 
-                       10 * arr_cnt)
-        arr_out = (arr_out 
-                   + np.multiply(arr_add,arr_out == 0).astype(np.uint8))
-        if test_mode: 
-            cv.imwrite('./debug/test_' + str(r) + '.jpg', 10 * arr_out)
-    mask = np.array(arr_out != 0).astype(int)
-    arr_out = np.multiply(arr_out - (np.max(arr_out)/2),mask)
-    return arr_out
-def plot_round(arr):
-    # function to smooth out the colors and crop out the borders
-    repeat = 10
-    arr_out = np.copy(arr)
-    # use queen smoothing to get rid of non-colored lines
-    for r in range(repeat):
-        arr_tmp = np.roll(
-            np.roll(
-                np.copy(arr_out), 
-                -1, 
-                axis = 0
-            ), 
-            -1, 
-            axis = 1
-        ).astype(np.float64)
-        # zero the edges
-        arr_tmp[0, :] = np.zeros_like(arr_tmp[0, :])
-        arr_tmp[-1,:] = np.zeros_like(arr_tmp[-1,:])
-        arr_tmp[:, 0] = np.zeros_like(arr_tmp[:, 0])
-        arr_tmp[:,-1] = np.zeros_like(arr_tmp[:,-1])
-        arr_add = np.zeros_like(arr_out).astype(np.float64)
-        arr_cnt = np.zeros_like(arr_out).astype(float)
-        # queen smoothing
-        for (i, j) in [(1,0),(1,0),(1,1),(1,1),(-1,0),(-1,0),(-1,1),(-1,1)]:
-            arr_tmp = np.roll(arr_tmp, i, axis = j)
-            arr_add = arr_add + arr_tmp
-            arr_cnt = (arr_cnt 
-                       + np.array(arr_tmp != 0).astype(type(arr_cnt[0,0])))
-        arr_cnt = arr_cnt * 6
-        arr_add = np.divide(
-            arr_add, 
-            arr_cnt,
-            out = np.zeros_like(arr_add), 
-            where=arr_cnt!=0
-        )
-        if test_mode: 
-            cv.imwrite('./debug/test_add_' + str(r) + '.jpg', 10 * arr_add)
-        if test_mode: 
-            cv.imwrite('./debug/test_cnt_' + str(r) + '.jpg', 10 * arr_cnt)
-        arr_out = (arr_out 
-                   + np.multiply(arr_add,arr_out == 0).astype(np.uint8))
-        if test_mode: 
-            cv.imwrite('./debug/test_' + str(r) + '.jpg', 10 * arr_out)
-    return arr_out
-def image_restore(plot,subject,dtg,scale):
-    image_template = cv.imread('./templates/' + subject + '/' + subject + "_template.jpg")
-
-    x, y = plot.shape
-    t, b, l, r = get_tblr(image_template)
-    empty = [0, 0, 124]
-    var = 50
-
-    image = np.zeros_like(image_template).astype(int)
-    image_bool = np.ones_like(image[:,:,0]).astype(bool)
-
-    if test_mode: cv.imwrite("./debug/test_restore_plot.jpg", plot * (256 / np.max(plot)))
-
-    for i in range(3): image_bool = np.multiply(image_bool, abs(image_template[:,:,i].astype(int) - empty[i]) < var)
-
-    # change the scalar plot to an RGB image
-    for i in range(len(scale))[1:]:
-        for j in range(3): image[t:b,l:r,j] = np.array(plot == i).astype(int) * scale[i, j]
-        if test_mode: cv.imwrite("./debug/test_restore_" + str(i) + ".jpg",image)
-
-    for i in range(3): 
-        image[:,:,i] = np.multiply(np.array(image_bool).astype(int),image[:,:,i]) + np.multiply(np.array(image_bool).astype(int), image_template[:,:,i])
-        if test_mode: cv.imwrite("./debug/test_restore_f" + str(i) + ".jpg",image)
-
-    # crop out a part of the image (namely the legend)
-    ##################################################
-    # config file: x1 = left most x, x2 = right most x
-    #              y1 = top most y,  y2 = bottom most y
-
-    # crop the image to remove the legend
-    if c.crop_y1 !=0: image[:c.crop_y1, :, :] =  image_template[:c.crop_y1, :, :].copy()
-    if c.crop_y2 !=0: image[c.crop_y2:, :, :] =  image_template[c.crop_y2:, :, :].copy()
-    if c.crop_x1 !=0: image[:, :c.crop_x1, :] =  image_template[:, :c.crop_x1, :].copy()
-    if c.crop_x2 !=0: image[:, c.crop_x2:, :] =  image_template[:, c.crop_x2:, :].copy()
-
-    image = cv.putText(image, subject + " - " + dtg, (c.crop_x1, c.crop_y2 + 5), cv.FONT_HERSHEY_SIMPLEX,
-                       .5, (0,0,0), 1, cv.LINE_AA, False)
-
-    return image
-
-
-#####################################################################
 # C o n v e r t   b a s e   1 0   < - >   b a s e   6 2
 #####################################################################
 # function to compress each coefficient into a single character from
 # a DFT array normalized to -1000 -> 1000
-def dec2aA1(i):    
-    if i < 10: return         str(i)
-    if i < 36: return         chr(i + ord('A') - 10)
-    return                    chr(i + ord('a') - 36)
-
-def aA12dec(s):
-    try:
-        if int(s) < 10: return int(s)
-    except:
-        None
-    if ord(s) < ord('a'): 
-        return                 int(ord(s) + 10 - ord('A'))
-    else: return               int(ord(s) + 36 - ord('a'))
-
-def coeff_round(x):
-    k_cr = 54
-    n_cr = 3
-    for i in range(n_cr):
-        for j in range(9):
-            if x > (9.5-j) * (10 ** (2 - i)): return k_cr
-            k_cr -= 1
-            if x < (j-9.5) * (10 ** (2 - i)): return k_cr
-            k_cr -= 1
-    return 0
-
-def coeff_unround(x):
-    if x == 0: return 0
-    k_cu = 1
-    n_cu = 3
-    for i in range(n_cu):
-        for j in range(2,11):
-            if x == k_cu: return (-1) * j * (10 ** i)
-            k_cu += 1
-            if x == k_cu: return        j * (10 ** i)
-            k_cu += 1
-    return 0
 
 def DFT_mapping(n):
     # save all coefficients to a single flattened array organized by:
